+++ vite-project/index.html
@@ -11,7 +11,7 @@
   <title>GSAP Project</title>
 </head>
 
-<body class="warm">
+<body id="body1" class="warm">
   <!-- Home Page -->
   <section class="homepage" id="Home">
     <div class="header">
